Use CBMC version 5.95.1
The upcoming CBMC version 6 release includes changes that may affect
existing proofs. This PR will make sure that
Fleet-Provisioning-for-AWS-IoT-embedded-sdk PRs are not negatively
impacted by this release. After releasing CBMC version 6 we will issue a
follow-up PR that will return
Fleet-Provisioning-for-AWS-IoT-embedded-sdk to using CBMC's latest
release, and will include any changes to proofs that may be necessary to
support the new version.